requirements.txt
gunicorn
Flask==2.0.0
lime==0.2.0.1
matplotlib==3.3.2
numpy==1.22.4
pandas==1.1.3
Pillow==9.4.0
scikit_learn==1.2.2
